Have entry deletion be a flag passed to the dataset view
